chore(gulpfile): style and feat
* reindent pipes
* add npm install in the end of release
* hook cb to release task

+++ gulpfile.ls
@@ -14,46 +14,46 @@ require! {
 
 gulp.task 'vendor' ->
   return gulp.src 'vendor/**/*'
-    .pipe gulp.dest 'public'
+  .pipe gulp.dest 'public'
 /*
  * dev subtasks
  */
 gulp.task 'dev:html' ->
   return gulp.src 'app/*.html'
-    .pipe gulp.dest 'public'
-    .pipe gulp-livereload(livereload)
+  .pipe gulp.dest 'public'
+  .pipe gulp-livereload(livereload)
 
 gulp.task 'dev:css' ->
   return gulp.src 'app/scss/application.scss'
-    .pipe gulp-exec('compass compile --force')
-    .pipe gulp-livereload(livereload)
+  .pipe gulp-exec('compass compile --force')
+  .pipe gulp-livereload(livereload)
 
 gulp.task 'dev:js' ->
   return gulp.src <[
-      bower_components/modernizr/modernizr.js
-      bower_components/jquery/dist/jquery.js
-      app/js/jquery.sticky.js
-      app/js/scrollit.min.js
-      bower_components/jquery.easy-pie-chart/dist/jquery.easypiechart.js
-      bower_components/foundation/js/foundation.js
-      app/js/app.js
-    ]>
-    .pipe gulp-concat 'application.js'
-    .pipe gulp.dest 'public'
-    .pipe gulp-livereload(livereload)
+    bower_components/modernizr/modernizr.js
+    bower_components/jquery/dist/jquery.js
+    app/js/jquery.sticky.js
+    app/js/scrollit.min.js
+    bower_components/jquery.easy-pie-chart/dist/jquery.easypiechart.js
+    bower_components/foundation/js/foundation.js
+    app/js/app.js
+  ]>
+  .pipe gulp-concat 'application.js'
+  .pipe gulp.dest 'public'
+  .pipe gulp-livereload(livereload)
 
 gulp.task 'dev:js:schedule' ->
   return gulp.src <[
-      bower_components/modernizr/modernizr.js
-      bower_components/jquery/dist/jquery.js
-      app/js/jquery.sticky.js
-      app/js/scrollit.min.js
-      bower_components/foundation/js/foundation.js
-      app/js/app-s.js
-    ]>
-    .pipe gulp-concat 'sys.js'
-    .pipe gulp.dest 'public'
-    .pipe gulp-livereload(livereload)
+    bower_components/modernizr/modernizr.js
+    bower_components/jquery/dist/jquery.js
+    app/js/jquery.sticky.js
+    app/js/scrollit.min.js
+    bower_components/foundation/js/foundation.js
+    app/js/app-s.js
+  ]>
+  .pipe gulp-concat 'sys.js'
+  .pipe gulp.dest 'public'
+  .pipe gulp-livereload(livereload)
 
 
 /*
@@ -61,29 +61,29 @@ gulp.task 'dev:js:schedule' ->
  */
 gulp.task 'public:css' ->
   return gulp.src 'app/scss/application.scss'
-    .pipe gulp-exec('compass compile --output-style compressed --force')
+  .pipe gulp-exec('compass compile --output-style compressed --force')
 
 gulp.task 'public:uglify' ->
   return gulp.src <[
-      bower_components/modernizr/modernizr.js
-      app/js/jquery.sticky.js
-      app/js/app.js
-    ]>
-    .pipe gulp-uglify!
-    .pipe gulp.dest 'tmp'
+    bower_components/modernizr/modernizr.js
+    app/js/jquery.sticky.js
+    app/js/app.js
+  ]>
+  .pipe gulp-uglify!
+  .pipe gulp.dest 'tmp'
 
 gulp.task 'public:js' <[ public:uglify ]> ->
   return gulp.src <[
-      tmp/modernizr.js
-      bower_components/jquery/dist/jquery.min.js
-      tmp/jquery.sticky.js
-      app/js/scrollit.min.js
-      bower_components/jquery.easy-pie-chart/dist/jquery.easypiechart.min.js
-      bower_components/foundation/js/foundation.min.js
-      tmp/app.js
-    ]>
-    .pipe gulp-concat 'application.js'
-    .pipe gulp.dest 'public'
+    tmp/modernizr.js
+    bower_components/jquery/dist/jquery.min.js
+    tmp/jquery.sticky.js
+    app/js/scrollit.min.js
+    bower_components/jquery.easy-pie-chart/dist/jquery.easypiechart.min.js
+    bower_components/foundation/js/foundation.min.js
+    tmp/app.js
+  ]>
+  .pipe gulp-concat 'application.js'
+  .pipe gulp.dest 'public'
 /*
  * for npm scripts
  */
@@ -108,14 +108,17 @@ gulp.task 'public' buildPublicSubtasks, !->
 
 gulp.task 'release' buildPublicSubtasks, !(cb) ->
   (err, dirpath) <-! temp.mkdir 'ntu-ccsp.github.io'
+  return cb err if err
   gulp.src 'package.json'
-    .pipe gulp-exec "cp -r public/* #{ dirpath }"
-    .pipe gulp-exec 'git checkout master'
-    .pipe gulp-exec 'git clean -f -d'
-    .pipe gulp-exec 'git rm -rf .'
-    .pipe gulp-exec "cp -r #{ path.join dirpath, '*' } ."
-    .pipe gulp-exec "rm -rf #{ dirpath }"
-    .pipe gulp-exec 'git add -A'
-    .pipe gulp-exec "git commit -m 'chore(release): by gulpfile'"
-    .pipe gulp-exec "git push origin master --force"
-    .pipe gulp-exec "git checkout dev"
+  .pipe gulp-exec "cp -r public/* #{ dirpath }"
+  .pipe gulp-exec 'git checkout master'
+  .pipe gulp-exec 'git clean -f -d'
+  .pipe gulp-exec 'git rm -rf .'
+  .pipe gulp-exec "cp -r #{ path.join dirpath, '*' } ."
+  .pipe gulp-exec "rm -rf #{ dirpath }"
+  .pipe gulp-exec 'git add -A'
+  .pipe gulp-exec "git commit -m 'chore(release): by gulpfile'"
+  .pipe gulp-exec "git push origin master --force"
+  .pipe gulp-exec "git checkout dev"
+  .pipe gulp-exec "npm install"
+  .on 'end' cb
